Teacher grade assistant
Project Description: You work for a company that does educational consulting. Your boss knows that you are a genius when it comes to Excel and decides to impose upon you a little – he assigns you a project to create an Excel spreadsheet that will be a teacher’s companion. It has to have an input area for the grades, rank them, split them into different quartiles, create relevant charts, give all the summary statistics that are relevant, color code grades that need to be looked into, have a conditional format highlighting when a student does better or worse than previous exams, show relevant lookup values for grades according to number and letter, have relevant pivot tables, etc. The goal is that it is a stand-alone tool for teachers to use to assess everything possible about a student.